renamed maproom library
